Finished the code for the conversion from python to java. tested I2C with sensor device.
