+++ weather/adafruit_bme680/src/test/java/com/adafruit/bme680/TestBME680.java
@@ -3,13 +3,14 @@ package com.adafruit.bme680;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.DisplayName;
 
+import com.adafruit.bme680.BME680;
 // import org.springframework.boot.test.context.SpringBootTest;
 
 public class TestBME680 {
 
     @Test
     @DisplayName("jmaksuta: test")
-    public void testBME680 {
+    public void testBME680() {
         BME680 bme680 = new BME680();
     }
